Update commands layout.
