+++ cmd/az.go
@@ -1,17 +1,3 @@
-/*
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
 package cmd
 
 import (
@@ -21,11 +7,10 @@ import (
 // azCmd represents the az command
 var azCmd = &cobra.Command{
 	Use:   "az",
-	Short: "Enable access to set of commands used to work with Azure cloud",
-	Long: `Enable access to set of commands used to work with Azure cloud:
-	- authentication - let you access authentication options - e.g. create Service Principal`,
+	Short: "Commands used to work with Azure cloud.",
+	Long:  `Commands used to work with Azure cloud.`,
 	PreRun: func(cmd *cobra.Command, args []string) {
-		debug("az called")
+		debug("az pre run called")
 	},
 }
